requirements.txt
Django==2.1.5
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sqlalchemy-clickhouse==0.1.5.post0
joblib==0.13.1
gunicorn==19.9.0
